Converted everything over to use glm::vec3 and floats

+++ src/ray.hpp
@@ -1,16 +1,16 @@
 
 #pragma once
 
-#include "vec3.hpp"
+#include <vec.hpp>
 
 struct ray {
   point3 orig;
   vec3 dir;
-  double time = 0.0;
+  float time = 0.0;
 
   constexpr ray() = default;
-  constexpr ray(const point3 &origin, const vec3 &direction, const double time)
+  constexpr ray(const point3 &origin, const vec3 &direction, const float time)
       : orig(origin), dir(direction), time(time) {}
 
-  constexpr point3 at(const double t) const { return orig + t * dir; }
+  constexpr point3 at(const float t) const { return orig + t * dir; }
 };
